+++ docs/changelog.rst
@@ -41,6 +41,7 @@ Fixes
 * :class:`aspecd.analysis.AggregateAnalysisStep` no longer adds ``datasets`` and ``result`` to output of ``to_dict()``
 * :class:`aspecd.tasks.AggregatedAnalysisTask` sets correct type in output of ``to_dict()``
 * :class:`aspecd.tasks.ReportTask` does not add empty figure filenames to includes
+* :class:`aspecd.Tasks.PlotTask` preserves labels of drawings
 
 
 Version 0.6.2